requirements.txt
fastapi
uvicorn
jinja2
google-generativeai
faster-whisper
ffmpeg-python
gTTS
python-multipart
supabase